# src/main/java/com/facts/financial_facts_service/entities/discount/models/trailingPriceData/TtmPriceData.java
package com.facts.financial_facts_service.entities.discount.models.trailingPriceData;

import jakarta.persistence.Entity;
import jakarta.persistence.Table;

@Entity
@Table(name = "ttm_price_data")
public class TtmPriceData extends AbstractTrailingPriceData {
}